# Backend/src/4Wheeler-swagger.yaml
openapi: 3.0.0
info:
  title: 4Wheeler API Documentation
  version: 1.0.0
  description: |-
    API documentation for the 4Wheeler Website.
    This project provides the backend services for managing vehicles for a car dealership, 
    helping customers to have a better look before deciding to go to the store and make a purchase.
    The application offers features such as user authentication, search and sort based on criteria functionality.
    The application is built with scalability, security, and modern REST API design principles in mind.
    Some useful links:
    - [The 4Wheeler_BE repository](https://github.com/Jung2304/4Wheeler_BE)
    - [The source API documentation for the 4Wheeler](https://github.com/Jung2304/4Wheeler_BE/blob/main/Backend/src/4Wheeler-swagger.yaml)
  contact:
    email: 23520333@gm.uit.edu.vn
servers:
  - url: http://localhost:4000
    description: Local server
  - url: https://fourwheeler.onrender.com/api-docs/
    description: Production server
...
paths:
...
  /api/cars/listing:
    get:
      tags:
        - cars
      summary: Get all cars with pagination, search, and category filters
      operationId: getCars
      parameters:
        - name: page
          in: query
          description: Page number for pagination (default is 1)
          required: false
          schema:
            type: integer
            example: 1
        - name: limit
          in: query
          description: Number of cars per page (default is 12)
          required: false
          schema:
            type: integer
            example: 12
        - name: search
          in: query
          description: Search by car make or model name (case-insensitive)
          required: false
          schema:
            type: string
            example: "BMW"
        - name: category
          in: query
          description: Filter by car category (e.g., SUV, Sedan, Hatchback)
          required: false
          schema:
            type: string
            example: "SUV"
      responses:
        '200':
          description: List of cars retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  page:
                    type: integer
                    example: 1
                  pages:
                    type: integer
                    example: 5
                  total:
                    type: integer
                    example: 50
                  cars:
                    type: array
                    items:
                      $ref: '#/components/schemas/CarResponse'
        '500':
          description: Server error while fetching cars
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Failed to fetch cars"
      security: []
  /api/cars/{id}:
    get:
      tags:
        - cars
      summary: Get car details by ID
      operationId: getCarDetail
      parameters:
        - name: id
          in: path
          description: Car ID (MongoDB ObjectId)
          required: true
          schema:
            type: string
            example: "65320a8c32f4e1b2c3d4e5f6"
      responses:
        '200':
          description: Car details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarResponse'
        '404':
          description: Car not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Car not found"
        '500':
          description: Server error while fetching car details
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Failed to fetch car details"
      security: []
  /api/cars/user/favorites:
    get:
      tags:
        - cars
      summary: Get user's favorite cars
      operationId: getUserFavorites
      description: Retrieve all cars saved to the user's favorites/wishlist
      responses:
        '200':
          description: Favorite cars retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                    example: 3
                  favorites:
                    type: array
                    items:
                      $ref: '#/components/schemas/CarResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User not found"
        '500':
          description: Server error while fetching favorites
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Failed to fetch favorites"
      security:
        - BearerAuth: []
  /api/cars/favorites/{carId}:
    post:
      tags:
        - cars
      summary: Add car to user's favorites
      operationId: addToFavorites
      description: Add a specific car to the user's favorites/wishlist
      parameters:
        - name: carId
          in: path
          description: Car ID to add to favorites
          required: true
          schema:
            type: string
            example: "65320a8c32f4e1b2c3d4e5f6"
      responses:
        '200':
          description: Car added to favorites successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Car added to favorites"
                  favorites:
                    type: array
                    items:
                      $ref: '#/components/schemas/CarResponse'
        '404':
          description: Car not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Car not found"
        '500':
          description: Server error while adding to favorites
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Failed to add to favorites"
      security:
        - BearerAuth: []

    delete:
      tags:
        - cars
      summary: Remove car from user's favorites
      operationId: removeFromFavorites
      description: Remove a specific car from the user's favorites/wishlist
      parameters:
        - name: carId
          in: path
          description: Car ID to remove from favorites
          required: true
          schema:
            type: string
            example: "65320a8c32f4e1b2c3d4e5f6"
      responses:
        '200':
          description: Car removed from favorites successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Car removed from favorites"
                  favorites:
                    type: array
                    items:
                      $ref: '#/components/schemas/CarResponse'
        '500':
          description: Server error while removing from favorites
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Failed to remove from favorites"
      security:
        - BearerAuth: []
  /api/cars/test-drive/{carId}:
    post:
      tags:
        - cars
      summary: Book a test drive for a specific car
      operationId: bookTestDrive
      description: Allows authenticated users to book a test drive. Email is automatically filled from the user's JWT token. A confirmation email is sent with booking details. Test drive is scheduled 2 days from booking date.
      parameters:
        - name: carId
          in: path
          description: The unique ID of the car to book test drive for
          required: true
          schema:
            type: string
            example: "65320a8c32f4e1b2c3d4e5f6"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - phone
              properties:
                name:
                  type: string
                  description: Full name of the person booking the test drive
                  example: "John Doe"
                phone:
                  type: string
                  description: Contact phone number
                  example: "+1234567890"
                preferredDate:
                  type: string
                  format: date
                  description: Preferred test drive date (optional)
                  example: "2025-12-15"
                message:
                  type: string
                  description: Additional message or special requests (optional)
                  example: "I'm interested in the premium package"
      responses:
        '201':
          description: Test drive booked successfully and confirmation email sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Test drive booked successfully! Confirmation email sent."
                  booking:
                    type: object
                    properties:
                      id:
                        type: string
                        example: "65f7b8c9d2e3a4f5b6c7d8e9"
                      carName:
                        type: string
                        example: "BMW X5"
                      testDriveDate:
                        type: string
                        example: "Wednesday, December 11, 2025"
                      email:
                        type: string
                        example: "user@example.com"
        '400':
          description: Missing required fields (name or phone)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Name and phone number are required!"
        '404':
          description: User or car not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Car not found!"
        '500':
          description: Server error while booking test drive
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Failed to book test drive!"
      security:
        - BearerAuth: []
...
  /api/admin/cars:
    get:
      tags:
        - admin
      summary: Get all cars including deleted (admin only)
      operationId: getAllCars
      parameters:
        - name: page
          in: query
          description: Page number for pagination (default is 1)
          required: false
          schema:
            type: integer
            example: 1
        - name: limit
          in: query
          description: Number of cars per page (default is 12)
          required: false
          schema:
            type: integer
            example: 12
        - name: search
          in: query
          description: Search by car make or model name (case-insensitive)
          required: false
          schema:
            type: string
            example: "BMW"
        - name: category
          in: query
          description: Filter by car category (e.g., SUV, Sedan, Hatchback)
          required: false
          schema:
            type: string
            example: "SUV"
      responses:
        '200':
          description: List of all cars retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  page:
                    type: integer
                    example: 1
                  pages:
                    type: integer
                    example: 5
                  total:
                    type: integer
                    example: 50
                  cars:
                    type: array
                    items:
                      $ref: '#/components/schemas/CarResponse'
        '500':
          description: Server error while fetching cars
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Failed to fetch cars due to server error!"
      security:
        - BearerAuth: []
  /api/admin/cars/{id}:
    get:
      tags:
        - admin
      summary: Get car details by ID including deleted (admin only)
      operationId: getCarDetailAdmin
      parameters:
        - name: id
          in: path
          description: Car ID (MongoDB ObjectId)
          required: true
          schema:
            type: string
            example: "65320a8c32f4e1b2c3d4e5f6"
      responses:
        '200':
          description: Car details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarResponse'
        '404':
          description: Car not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Car not found"
        '500':
          description: Server error while fetching car details
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Failed to fetch car details"
      security:
        - BearerAuth: []

    put:
      tags:
        - admin
      summary: Update car details (admin only)
      operationId: updateCar
      parameters:
        - name: id
          in: path
          description: Car ID (MongoDB ObjectId)
          required: true
          schema:
            type: string
            example: "65320a8c32f4e1b2c3d4e5f6"
      requestBody:
        description: Car details to update (all fields optional)
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                make:
                  type: string
                  example: "BMW"
                model:
                  type: string
                  example: "X5"
                year:
                  type: integer
                  example: 2023
                price:
                  type: number
                  format: double
                  example: 75000
                color:
                  type: string
                  example: "Black"
                category:
                  type: string
                  example: "SUV"
                seats:
                  type: integer
                  example: 5
                transmission:
                  type: string
                  example: "Automatic"
                fuelType:
                  type: string
                  example: "Gasoline"
                engine:
                  type: string
                  example: "2.0L Turbo"
                horsepower:
                  type: integer
                  example: 335
                images:
                  type: array
                  items:
                    type: string
                  example: ["image1.jpg"]
                description:
                  type: string
                  example: "Updated description"
      responses:
        '200':
          description: Car updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Car updated successfully!"
                  car:
                    $ref: '#/components/schemas/CarResponse'
        '404':
          description: Car not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Car not found"
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Failed to update car!"
      security:
        - BearerAuth: []

    delete:
      tags:
        - admin
      summary: Delete car (soft delete - admin only)
      operationId: deleteCar
      parameters:
        - name: id
          in: path
          description: Car ID (MongoDB ObjectId)
          required: true
          schema:
            type: string
            example: "65320a8c32f4e1b2c3d4e5f6"
      responses:
        '200':
          description: Car deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Car deleted successfully!"
        '404':
          description: Car not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Car not found"
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Failed to delete car!"
      security:
        - BearerAuth: []
...
components:
  schemas:
...
    CarResponse:
      type: object
      properties:
        _id:
          type: string
          example: "65320a8c32f4e1b2c3d4e5f6"
        make:
          type: string
          example: "BMW"
        model:
          type: string
          example: "X5"
        year:
          type: integer
          example: 2023
        price:
          type: number
          format: double
          example: 75000
        color:
          type: string
          example: "Black"
        category:
          type: string
          example: "SUV"
        seats:
          type: integer
          example: 5
        transmission:
          type: string
          example: "Automatic"
        fuelType:
          type: string
          example: "Gasoline"
        engine:
          type: string
          example: "2.0L Turbo"
        horsepower:
          type: integer
          example: 335
        images:
          type: array
          items:
            type: string
          example: ["image1.jpg", "image2.jpg"]
        description:
          type: string
          example: "Premium luxury SUV with advanced features"
        deleted:
          type: boolean
          example: false
        createdAt:
          type: string
          format: date-time
          example: "2023-10-25T12:34:56.789Z"
        updatedAt:
          type: string
          format: date-time
          example: "2023-10-25T12:34:56.789Z"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Use the `Bearer` scheme to authenticate requests. Include the token in the `Authorization` header as follows:
        ```
        Authorization: Bearer <your_token>
        ```
security:
  - BearerAuth: []